Exercise 4 DataManipulation
Exercise containing two dictionaries. One contains people's names as a key and a dictionary as a value, which contains the cities where the shipment will take place and the weight of the package.

Then, there is another dictionary where the cities and the total weight of the shipments made to that city will be stored.

Finally, through a drop-down menu, the user will choose a city and the name of the city and the total weight of the shipment will be displayed on the screen.