Example Exercise [Excel & DataTables]
Test exercise showing how the DataTable works.

Two DataTables are created containing user and overdue book loans.

Subsequently, a union of the two DataTable is made and the following formats are applied to them: Delete unnecessary columns and ordering of the data.

Lastly, the data is displayed on the console.