Example_2_Excel-DataTables
In which two files are used, one with an .xlsx extension and the other with a .xls extension. Both contain company information.

Two data tables are used to obtain information from both files and finally, through the use of conditions, we filter so that we only obtain companies that were founded before 2005.

Finally, we create a third DataTable that will contain the filtered data and we will save it in an .xlsx file