Example 3 Excel-DataTables
Exercise in which an excel file with an .xlsx extension containing employees with different fields is read.

Data filtering is performed with two conditions.

1) People under 30 years old OR
2) People with incomes over 100,000

A DataTable is defined with three columns (Age, Name and Income).

Then, by means of a For Each Row, the DataTable in which the data was initially dumped is traversed, the name, age and income are obtained for each row and each one is stored in generic variables. And they will be added to the DataTable created previously.

Finally, the data contained in the DataTable that was created is dumped into an .xlsx file