Exercise-1_Excel-DataTables
Exercise in which the DataTable is used to read a file in .xlsx format from which we obtain the data in columns A and B. To later perform the sum in Column C.

The exercise is done in three ways. In one of them, the Excel Application Scope is not used, and in two others an Excel Application Scope is used.

Also, in two of the modes the addition process is carried out through a For Each Row. Instead there is one in which the SUM formula is used to perform the sum in a given range of rows.