Exercise-3_Excel-DataTables
Exercise in which two .xlsx files are provided containing information on expenses incurred with a card and in cash.
The idea is to bring the data from the two files together and display it in a single file.
You have to obtain a percentage of the expenses incurred in each category.