Input Actions and Input Method OCR Example 5
Exercise in which methods are used to work with the elements of the UI. For example to search for specific elements or check if a certain element exists.

A process is automated that allows you to log in to GMail and obtain the number of emails that the user has in his account.

Then the user is indicated the number of emails through an alert and a message in the console.

Finally, the session is closed and the user's account is deleted from the browser.