Exercise 1 Input Actions and Input Method
Exercise in which the user is asked to enter a text and file name.

The process of opening the notepad, writing the text that the user has indicated, changing the format and accessing the Save as option to save the file with the name that the user has specified is automated.