Exercise 1 Project Organization
Workflow that simulates a beverage vending machine.

The user chooses a starting balance and then the available drinks are displayed.

As the user purchases items, the value of the product will be subtracted from their balance.

You are also allowed the option to refund the amount of the last product purchased.

If the user does not have enough balance or chooses to complete the purchase, the process will end.