Exercise 2 Project Organization
Exercise in which a workflow is extracted for reuse.

The extracted flow is responsible for accessing a website that generates information about people, processes it and returns their personal data.

Then, the main flow through a while, adds information from five people to a datatable to later save it in a .xlsx file