Exercise 1 Error And Exception Handling
Exercise that reads an .xlsx file and verifies if the data contained in the Cash In and Cash Out row is correct (when converting from String to Int32). If the conversion is correct, the difference between Cash In, Cash Out is calculated and the status is marked as Success.

If, on the contrary, any of the Cash In or Cash Out data cannot be converted, it is shown which of the two or both could not be processed due to being wrong.

Finally, the calculations are stored in an .xlsx file